transformers_daemon/requirements.txt
# Transformers Autonomous Daemon Requirements

# Core dependencies
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0

# Web API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Async support
aiofiles>=23.0.0
asyncio>=3.4.3

# Crypto and Bitcoin
requests>=2.31.0
websockets>=11.0

# Optional: File monitoring (Linux)
watchdog>=3.0.0

# Optional: Database for memory
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Optional: Advanced AI features
# openai>=1.0.0  # Uncomment if using OpenAI integration
# anthropic>=0.3.0  # Uncomment if using Claude integration

# Development tools (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
